[spotify] Reorganise in preparation of new Spotify integration
