+++ src/artwork.c
@@ -46,7 +46,7 @@
 #include "artwork.h"
 
 #ifdef HAVE_SPOTIFY_H
-# include "spotify_webapi.h"
+# include "library/spotify_webapi.h"
 #endif
 
 /* This artwork module will look for artwork by consulting a set of sources one
