+++ src/inputs/spotify.c
@@ -25,59 +25,22 @@
 #include "logger.h"
 #include "spotify.h"
 
-// How many retries to start playback if resource is still loading
-#define SPOTIFY_SETUP_RETRIES 5
-// How long to wait between retries in microseconds (500000 = 0.5 seconds)
-#define SPOTIFY_SETUP_RETRY_WAIT 500000
-
 static int
 setup(struct input_source *source)
 {
-  int i = 0;
-  int ret;
-
-  while((ret = spotify_playback_setup(source->path)) == SPOTIFY_SETUP_ERROR_IS_LOADING)
-    {
-      if (i >= SPOTIFY_SETUP_RETRIES)
-	break;
-
-      DPRINTF(E_DBG, L_SPOTIFY, "Resource still loading (%d)\n", i);
-      usleep(SPOTIFY_SETUP_RETRY_WAIT);
-      i++;
-    }
-
-  if (ret < 0)
-    return -1;
-
-  ret = spotify_playback_play();
-  if (ret < 0)
-    return -1;
-
-  return 0;
+  return -1;
 }
 
 static int
 stop(struct input_source *source)
 {
-  int ret;
-
-  ret = spotify_playback_stop();
-  if (ret < 0)
-    return -1;
-
-  return 0;
+  return -1;
 }
 
 static int
 seek(struct input_source *source, int seek_ms)
 {
-  int ret;
-
-  ret = spotify_playback_seek(seek_ms);
-  if (ret < 0)
-    return -1;
-
-  return ret;
+  return -1;
 }
 
 struct input_definition input_spotify =
@@ -90,3 +53,30 @@ struct input_definition input_spotify =
   .seek = seek,
 };
 
+
+/* ------------ Functions exposed via spotify.h, mocked for now ------------- */
+// Th
+int
+spotify_login_user(const char *user, const char *password, char **errmsg)
+{
+  return -1;
+}
+
+void
+spotify_login(char **arglist)
+{
+  return;
+}
+
+
+void
+spotify_logout(void)
+{
+  return;
+}
+
+void
+spotify_status_info_get(struct spotify_status_info *info)
+{
+  return;
+}
